db연동 ver2

+++ app/src/main/java/com/example/arnavi/Data/BuildingDTO.java
@@ -3,8 +3,8 @@ package com.example.arnavi.Data;
 public class BuildingDTO {
 
     private String bname;    //건물이름
-    private float latitude;     //위도
-    private float longtitude;   //경도
+    private double latitude;     //위도
+    private double longtitude;   //경도
 
     public String getBname() {
         return bname;
@@ -14,19 +14,19 @@ public class BuildingDTO {
         this.bname = bname;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongtitude() {
+    public double getLongtitude() {
         return longtitude;
     }
 
-    public void setLongtitude(float longtitude) {
+    public void setLongtitude(double longtitude) {
         this.longtitude = longtitude;
     }
 }
